backend/ghaminer1/src/tmp/oxygencs-grp1-eq4/requirements.txt
pytest
python-dotenv
signalrcore
psycopg2
